Spotify (test) animation added (UNLICENSED)
You can do whatever you want with the animation, it was just a test. 

I do not own nor do I'm in any association with Spotify AB which means I have no rights on their Logo and there before also no right to license it even in the case that I've created the animation.